create disclosure component
